requirements.txt
Django==5.0.6
djangorestframework==3.15.1
mysqlclient==2.2.4
Pillow==10.2.0
django-cors-headers==4.3.1
django-environ==0.4.5